+++ Localize/lang/strings.json
@@ -639,6 +639,7 @@
   "F1AkvV": "Ideal for process business transitional data",
   "F2wR+r": "Insert expression",
   "F3IDl8": "Required. The number of a specified time unit to add.",
+  "F83QRP": "On a schedule",
   "F9dR1Q": "Add",
   "FDF4Qb": "Returns a parameter value that is defined in the definition",
   "FEYdkx": "Maximum waiting runs",
@@ -717,6 +718,7 @@
   "HDqP2g": "Required. The key name of the form data value to return.",
   "HET2nV": "Add a Variable",
   "HF2SNx": "Is successful",
+  "HFt+tF": "Favorites",
   "HH970i": "Month",
   "HILmmE": "Required. The collection to sort.",
   "HMiE+4": "Exit full screen",
@@ -735,12 +737,12 @@
   "HoFPHB": "When something happens in Azure services",
   "HzS2gJ": "Dynamic content not supported as properties in authentication.",
   "I+85NV": "Submit from this action",
+  "I+jX2I": "Trigger manually",
   "I1CYNA": "Invalid property ''{invalidProperties}'' for authentication type ''{authType}''.",
   "I2XWRg": "Add",
   "I2Ztna": "Loop automatically added when connecting a repeating source element. No function required.",
   "I3mifR": "Is skipped",
   "I41vZ/": "(UTC-11:00) Coordinated Universal Time-11",
-  "I547J5": "On a Schedule",
   "IA+Ogm": "22",
   "IAmvpa": "(UTC-08:00) Coordinated Universal Time-08",
   "IBFBR2": "Remove loop",
@@ -1134,7 +1136,6 @@
   "SyFXM3": "Read less",
   "Sz8KN3": "Test",
   "T/7b2y": "Duration",
-  "T1JBCn": "Trigger Manually",
   "T1q9LE": "Name",
   "TBagKD": "No operation selected",
   "TEN+cR": "Give feedback",
@@ -1264,7 +1265,6 @@
   "W+mUyI": "Next",
   "W070M2": "of {max}",
   "W1rlxU": "State type",
-  "W4TFz4": "No operations found for this category",
   "W621ZA": "Contact info",
   "W6FdMh": "Required. The name of the new property.",
   "W83QYZ": "No results for",
@@ -2069,6 +2069,7 @@
   "_F1AkvV.comment": "Second bullet point of stateful type",
   "_F2wR+r.comment": "Label for button to open expression token picker",
   "_F3IDl8.comment": "Required integer parameter to add to time",
+  "_F83QRP.comment": "Schedule trigger category",
   "_F9dR1Q.comment": "Add",
   "_FDF4Qb.comment": "Label for description of custom parameters Function",
   "_FEYdkx.comment": "Label for maximum waiting runs",
@@ -2147,6 +2148,7 @@
   "_HDqP2g.comment": "Required string parameter to be used as key for triggerFormDataValue function",
   "_HET2nV.comment": "label to add a variable",
   "_HF2SNx.comment": "Successful action result",
+  "_HFt+tF.comment": "Favorites category",
   "_HH970i.comment": "Frequency value ",
   "_HILmmE.comment": "Required collection parameter to apply sort function on",
   "_HMiE+4.comment": "Token picker for 'Exit full screen'",
@@ -2165,12 +2167,12 @@
   "_HoFPHB.comment": "Azure events trigger category description",
   "_HzS2gJ.comment": "Error message for when putting token in authentication property",
   "_I+85NV.comment": "Button label for submitting a workflow to rerun from this action",
+  "_I+jX2I.comment": "Manual trigger category",
   "_I1CYNA.comment": "Error message when having an invalid authentication property",
   "_I2XWRg.comment": "Button text for opening panel for adding workflows",
   "_I2Ztna.comment": "Message explaining user does not need to add a loop function",
   "_I3mifR.comment": "Skipped run",
   "_I41vZ/.comment": "Time zone value ",
-  "_I547J5.comment": "Schedule trigger category",
   "_IA+Ogm.comment": "Hour of the day",
   "_IAmvpa.comment": "Time zone value ",
   "_IBFBR2.comment": "Remove loop for the connection",
@@ -2564,7 +2566,6 @@
   "_SyFXM3.comment": "Text for read less button",
   "_Sz8KN3.comment": "Test",
   "_T/7b2y.comment": "Duration column header",
-  "_T1JBCn.comment": "Manual trigger category",
   "_T1q9LE.comment": "The label for the connector column",
   "_TBagKD.comment": "Message displayed when no operation is selected in the edit operation panel",
   "_TEN+cR.comment": "Button text for submitting feedback",
@@ -2694,7 +2695,6 @@
   "_W+mUyI.comment": "Placeholder text for the Next button in the suggested workflow description",
   "_W070M2.comment": "Text on a pager where people can select a page number out of {max}",
   "_W1rlxU.comment": "Label for choosing State type",
-  "_W4TFz4.comment": "No operations message",
   "_W621ZA.comment": "Contact info section title",
   "_W6FdMh.comment": "Required string parameter for new property name in addProperty function",
   "_W83QYZ.comment": "Text displayed when no results match the search term",
@@ -3123,6 +3123,7 @@
   "_h+W3VW.comment": "Label for Number type dynamically added parameter",
   "_h+ZYip.comment": "Option to install a new gateway, links to new page",
   "_h1lQDa.comment": "Modal Title text",
+  "_h2OdHF.comment": "Text for the Action Details header",
   "_h2erb6.comment": "Label for agent parameters",
   "_h3V6DM.comment": "Label for trigger description",
   "_h40rpg.comment": "description of retry interval setting",
@@ -4161,6 +4162,7 @@
   "h+W3VW": "Number",
   "h+ZYip": "{addIcon} Install gateway",
   "h1lQDa": "Enter or paste a sample JSON payload.",
+  "h2OdHF": "Action Details",
   "h2erb6": "Add agent parameters",
   "h3V6DM": "Trigger description",
   "h40rpg": "Specify interval in ISO 8601 format.",
