+++ apps/Standalone/package.json
@@ -9,7 +9,7 @@
     "@fluentui/react-components": "9.56.0",
     "@fluentui/react-focus": "8.8.19",
     "@fluentui/react-hooks": "8.6.20",
-		"@fluentui/react-icons": "2.0.224",
+    "@fluentui/react-icons": "2.0.224",
     "@fluentui/react-migration-v8-v9": "^9.6.6",
     "@fluentui/react-portal-compat": "^9.0.138",
     "@fluentui/style-utilities": "8.9.6",
