+++ libs/designer-v2/package.json
@@ -62,10 +62,7 @@
     },
     "./package.json": "./package.json"
   },
-  "files": [
-    "build/lib/**/*",
-    "src"
-  ],
+  "files": ["build/lib/**/*", "src"],
   "license": "MIT",
   "main": "src/index.ts",
   "module": "src/index.ts",
