+++ libs/designer-v2/src/lib/ui/panel/recommendation/browse/styles/CategoryCard.styles.ts
