+++ libs/designer-v2/src/lib/ui/panel/recommendation/browse/styles/ConnectorCard.styles.ts
@@ -1,0 +1,94 @@
+import { makeStyles, tokens } from '@fluentui/react-components';
+
+export const useConnectorCardStyles = makeStyles({
+  card: {
+    cursor: 'pointer',
+    display: 'flex',
+    alignItems: 'center',
+    borderRadius: tokens.borderRadiusMedium,
+    backgroundColor: tokens.colorNeutralBackground1,
+    position: 'relative',
+    padding: `${tokens.spacingVerticalM} ${tokens.spacingHorizontalL}`,
+    transition: 'all 0.2s ease',
+    border: `1px solid ${tokens.colorNeutralStroke2}`,
+    marginBottom: tokens.spacingVerticalS,
+    '&:hover': {
+      backgroundColor: tokens.colorNeutralBackground1Hover,
+      border: `1px solid ${tokens.colorNeutralStroke1Hover}`,
+      '&::before': {
+        opacity: 1,
+      },
+    },
+    '&::before': {
+      content: '""',
+      position: 'absolute',
+      left: 0,
+      top: 0,
+      bottom: 0,
+      width: '4px',
+      backgroundColor: tokens.colorBrandBackground,
+      borderTopLeftRadius: tokens.borderRadiusMedium,
+      borderBottomLeftRadius: tokens.borderRadiusMedium,
+      opacity: 0,
+      transition: 'opacity 0.2s ease',
+    },
+  },
+  iconContainer: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginRight: tokens.spacingHorizontalM,
+    width: '32px',
+    height: '32px',
+    flexShrink: 0,
+  },
+  icon: {
+    width: '24px',
+    height: '24px',
+    borderRadius: tokens.borderRadiusSmall,
+  },
+  content: {
+    flex: 1,
+    minWidth: 0,
+    display: 'flex',
+    flexDirection: 'column',
+  },
+  titleRow: {
+    display: 'flex',
+    alignItems: 'center',
+    marginBottom: '2px',
+  },
+  title: {
+    overflow: 'hidden',
+    textOverflow: 'ellipsis',
+    whiteSpace: 'nowrap',
+    fontWeight: tokens.fontWeightSemibold,
+    fontSize: tokens.fontSizeBase300,
+    color: tokens.colorNeutralForeground1,
+    lineHeight: tokens.lineHeightBase300,
+  },
+  builtInBadge: {
+    marginLeft: tokens.spacingHorizontalS,
+    padding: '2px 6px',
+    backgroundColor: tokens.colorBrandBackground2,
+    borderRadius: '10px',
+    fontSize: '10px',
+    fontWeight: tokens.fontWeightSemibold,
+    color: tokens.colorBrandForeground2,
+    textTransform: 'uppercase',
+    letterSpacing: '0.5px',
+  },
+  description: {
+    fontSize: tokens.fontSizeBase200,
+    color: tokens.colorNeutralForeground2,
+    lineHeight: tokens.lineHeightBase200,
+    overflow: 'hidden',
+    textOverflow: 'ellipsis',
+    whiteSpace: 'nowrap',
+  },
+  chevron: {
+    marginLeft: tokens.spacingHorizontalM,
+    flexShrink: 0,
+    color: tokens.colorNeutralForeground3,
+  },
+});
