+++ libs/designer-v2/src/lib/ui/panel/recommendation/details/operationAccordionItem.tsx
