+++ libs/designer-v2/src/lib/ui/panel/recommendation/details/styles/ConnectorDetailsView.styles.ts
