+++ libs/designer-v2/src/lib/ui/panel/recommendation/details/styles/OperationsAccordion.styles.ts
