--- libs/designer-v2/src/lib/ui/panel/recommendation/styles/TriggerOperationList.styles.ts
@@ -1,11 +1,0 @@
-import { makeStyles } from '@fluentui/react-components';
-
-export const useTriggerOperationListStyles = makeStyles({
-  operationIcon: {
-    width: '24px',
-    height: '24px',
-  },
-  noOperationsText: {
-    opacity: 0.7,
-  },
-});
