+++ package.json
@@ -16,6 +16,7 @@
     "@types/react": "18.3.0",
     "@types/react-dom": "18.3.0",
     "@types/react-test-renderer": "^18.0.7",
+    "@types/react-window": "^1.8.8",
     "@typescript-eslint/eslint-plugin": "^8.15.0",
     "@typescript-eslint/parser": "^8.29.1",
     "@vitejs/plugin-react": "^4.4.6",
@@ -67,7 +68,11 @@
   },
   "license": "MIT",
   "lint-staged": {
-    "*.{js,ts,tsx}": ["npm run extract", "eslint --cache --fix", "biome check --write"]
+    "*.{js,ts,tsx}": [
+      "npm run extract",
+      "eslint --cache --fix",
+      "biome check --write"
+    ]
   },
   "private": true,
   "scripts": {
@@ -102,6 +107,7 @@
     "find-process": "^1.4.7",
     "fs-extra": "^11.2.0",
     "happy-dom": "^18.0.1",
+    "react-window": "^1.8.11",
     "ts-node": "^10.9.2"
   },
   "pnpm": {
